feature:gpu: use double instead

+++ lib/app/features/gpu/model/gpu_model.dart
@@ -1,9 +1,13 @@
 class GpuModel {
   final String name;
   final String currentFrequencyNode;
-  final List<int> availableFrequencies;
+  final List<double> availableFrequencies;
   final List<String> availableGovernors;
 
-  GpuModel(this.name, this.currentFrequencyNode, this.availableFrequencies,
-      this.availableGovernors,);
+  GpuModel(
+    this.name,
+    this.currentFrequencyNode,
+    this.availableFrequencies,
+    this.availableGovernors,
+  );
 }
